projeto-djangoweaver/djangoProject/requirements.txt
PyPDF2>=3.0.1
reportlab~=4.2.0
Django>=5.0.6
xmltodict~=0.13.0
celery~=5.4.0
djangorestframework~=3.15.1
requests~=2.32.3
python-dotenv~=1.0.1
croniter~=3.0.3
requests-oauthlib~=2.0.0
msal~=1.31.0